requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
httpx>=0.27,<1
python-dotenv>=1,<2
transformers>=4.40,<5
torch>=2.2,<3
sqlalchemy>=2.0,<3
alembic>=1.13,<2
passlib[bcrypt]>=1.7,<2
bcrypt>=3.2,<4
python-jose[cryptography]>=3.3,<4
psycopg2-binary>=2.9,<3
email-validator>=2.1,<3

# Testing
pytest>=7.4,<8
pytest-asyncio>=0.23,<1
pytest-cov>=4.1,<5
faker>=22.0,<23
